Navigation system reformatted
