+++ components/Stylesheet.tsx
@@ -1,0 +1,19 @@
+import {ColorValue, FlexAlignType, FlexStyle} from "react-native";
+
+export interface StyleSheet {
+    container: {
+        flex: number;
+        backgroundColor: ColorValue;
+        color: ColorValue;
+        alignItems: FlexAlignType;
+        justifyContent: FlexStyle["justifyContent"];
+    },
+    text: {
+        color: ColorValue;
+        backgroundColor: ColorValue;
+    },
+    button: {
+        backgroundColor: ColorValue;
+        color: ColorValue;
+    }
+}
